requirements.txt
requests==2.32.3
python-dotenv==1.0.1
matplotlib==3.9.2
tenacity==9.0.0
pytest==8.3.2
flake8==7.1.1
pandas==2.2.3
Flask==3.0.3
gunicorn==21.2.0
